e2fsprogs: check if e2scrub_all_cron exists
doesn’t exist on macos

+++ pkgs/tools/filesystems/e2fsprogs/default.nix
@@ -39,7 +39,9 @@ stdenv.mkDerivation rec {
 
   postInstall = ''
     # avoid cycle between outputs
-    mv $out/lib/${pname}/e2scrub_all_cron $bin/bin/
+    if [ -f $out/lib/${pname}/e2scrub_all_cron ]; then
+      mv $out/lib/${pname}/e2scrub_all_cron $bin/bin/
+    fi
   '';
 
   enableParallelBuilding = true;
